+++ tensorflow_similarity/retrieval_metrics/precision_at_k.py
@@ -67,7 +67,7 @@ class PrecisionAtK(RetrievalMetric):
             embedding queries.
 
             match_mask: A 2D mask where a 1 indicates a match between the
-            jth query and the kth neighboor and a 0 indicates a mismatch.
+            jth query and the kth neighbor and a 0 indicates a mismatch.
 
             **kwargs: Additional compute args.